+++ Tests/Test_SetC_Utilities/test_a_performance/test_b_array/test_a_performance_array_package.py
@@ -1,0 +1,11 @@
+""" Test the `TidalPy.utilities.performance.array` module. """
+
+import TidalPy
+TidalPy.test_mode()
+
+
+def test_utilities_performance_array_package():
+    """ Test `utilities.performance.array` import. """
+
+    from TidalPy.utilities.performance.array.interp import interp
+    assert True