+++ TidalPy/utilities/performance/array/__init__.py
@@ -1,0 +1,2 @@
+# Import `interp` function
+from TidalPy.utilities.performance.array.interp import interp