+++ TidalPy/utilities/performance/special/__init__.py
@@ -1,0 +1,1 @@
+from .factorial import find_factorial